tests/wpt/mozilla/tests/bluetooth/requestDevice/blocklisted-service-in-filter.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(t => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.blocklist);
    return promise_rejects_dom(
        t, 'SecurityError', window.navigator.bluetooth.requestDevice({filters: [{services: [human_interface_device.name]}]}),
        'Requesting blocklisted service rejects.');
}, 'Reject with SecurityError if requesting a blocklisted service.');
</script>
